pyproject.toml
[tool.poetry]
name = "hammer-systems"
version = "0.1.0"
description = ""
authors = ["Ayta Dzhivanov <kurt.alive.1994@gmail.com>"]
readme = "README.md"
packages = [{include = "hammer_systems"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.4"
djangorestframework = "^3.14.0"
psycopg2 = "^2.9.7"
django-phonenumber-field = {extras = ["phonenumbers"], version = "^7.1.0"}
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
